Move float16 behind a flag

+++ cosmobot_deep_learning/prepare_dataset.py
@@ -4,7 +4,6 @@ from typing import Dict
 import numpy as np
 import pandas as pd
 
-
 from cosmobot_deep_learning.preprocess_image import (
     open_and_preprocess_images,
     open_and_preprocess_image_ROIs,
@@ -27,7 +26,7 @@ def extract_inputs(df, input_column_names):
         **{"sr": df["DO patch r_msorm"] / df["reference patch r_msorm"]}
     )[input_column_names]
 
-    return dataset.values.astype("float16")
+    return dataset.values
 
 
 def extract_labels(df, label_column, label_scale_factor_mmhg):
@@ -43,7 +42,7 @@ def extract_labels(df, label_column, label_scale_factor_mmhg):
     scaled_labels = df[label_column] / label_scale_factor_mmhg
 
     # Reshape to 2d array
-    return np.reshape(scaled_labels.values, (-1, 1)).astype("float16")
+    return np.reshape(scaled_labels.values, (-1, 1))
 
 
 def prepare_dataset_numeric(raw_dataset: pd.DataFrame, hyperparameters):
